refactor(peagen): move JSON schemas package
