+++ pkgs/standards/peagen/docs/call_flows/peagen_extra-schemas_generate_remote.mmd
@@ -4,12 +4,12 @@ sequenceDiagram
     participant FS as File System
     User ->> CLI: Run `peagen remote extras-schemas generate`
     CLI ->> FS: Determine scaffold templates directory and output schemas directory
-    CLI ->> FS: Create/ensure `schemas/extras/` directory exists
+    CLI ->> FS: Create/ensure `jsonschemas/extras/` directory exists
     CLI ->> FS: Find all `EXTRAS.md` files in templates subfolders
     loop for each EXTRAS.md file found
         CLI ->> FS: Read the markdown file
         CLI ->> CLI: Parse lines for keys (bulleted list items)
         CLI ->> CLI: Build JSON schema object with those keys as properties
-        CLI ->> FS: Write schema JSON to `schemas/extras/<set_name>.schema.v1.json`
+        CLI ->> FS: Write schema JSON to `jsonschemas/extras/<set_name>.schema.v1.json`
         CLI -->> User: Print "✅ Wrote <set_name>.schema.v1.json"
     end
