+++ pkgs/standards/peagen/docs/pea-ips/ip-0004.md
@@ -144,7 +144,7 @@ If the same key appears multiple times the **earlier** entry wins, ensuring CLI 
 | Workspace loader | `program.py` / `common.temp_workspace`                            | Re-use fetch logic when URI is remote.            |
 | Default Pool     | `peagen/eval/pools/default.py`                                        | Threaded pool honouring `max_workers` & `async`.  |
 | Record writer   | `record_writer.py`                                                   | Add `mode="eval"` + JSONL roll-up.                |
-| Schema files     | `schemas/eval_report.schema.v1.json`, update `peagen.toml.schema.*` | Author & unit-test with `ajv`.                    |
+| Schema files     | `jsonschemas/eval_report.schema.v1.json`, update `peagen.toml.schema.*` | Author & unit-test with `ajv`.                    |
 | Tests            | `tests/eval/test_eval_cmd.py`                                         | Functional coverage incl. strict-mode exit codes. |
 | Docs             | `docs/cli/eval.md`                                                    | End-to-end guide, CI snippet.                     |
