+++ pkgs/standards/peagen/peagen/jsonschemas/doe_spec.schema.v1.1.json
