+++ pkgs/standards/peagen/peagen/jsonschemas/doe_spec.schema.v1.json
