+++ pkgs/standards/peagen/peagen/jsonschemas/doe_spec.schema.v2.json
