+++ pkgs/standards/peagen/peagen/jsonschemas/event.schema.v1.json
