+++ pkgs/standards/peagen/peagen/jsonschemas/extras/component.schema.v1.json
