+++ pkgs/standards/peagen/peagen/jsonschemas/extras/cpp_python_pkg.schema.v1.json
