+++ pkgs/standards/peagen/peagen/jsonschemas/extras/python_api_gw_service.schema.v1.json
