+++ pkgs/standards/peagen/peagen/jsonschemas/extras/pythonpkg.schema.v1.json
