+++ pkgs/standards/peagen/peagen/jsonschemas/extras/react_atom.schema.v1.json
