+++ pkgs/standards/peagen/peagen/jsonschemas/extras/rust_python_pkg.schema.v1.json
