+++ pkgs/standards/peagen/peagen/jsonschemas/extras/svelte_atom.schema.v1.json
