+++ pkgs/standards/peagen/peagen/jsonschemas/extras/swarmauri_base.schema.v1.json
