+++ pkgs/standards/peagen/peagen/jsonschemas/extras/swarmauri_core.schema.v1.json
