+++ pkgs/standards/peagen/peagen/jsonschemas/extras/swarmauri_standard.schema.v1.json
