+++ pkgs/standards/peagen/peagen/jsonschemas/extras/swarmauri_standard_standalone.schema.v1.json
