+++ pkgs/standards/peagen/peagen/jsonschemas/extras/test3.schema.v1.json
