+++ pkgs/standards/peagen/peagen/jsonschemas/extras/vue_atom.schema.v1.json
