+++ pkgs/standards/peagen/peagen/jsonschemas/manifest.schema.v3.1.json
