+++ pkgs/standards/peagen/peagen/jsonschemas/projects_payload.schema.v1.json
