+++ pkgs/standards/peagen/peagen/jsonschemas/ptree.schema.v1.json
