+++ pkgs/standards/peagen/pyproject.toml
@@ -206,7 +206,7 @@ simple_time = "peagen.plugins.evaluators.simple_time:SimpleTimeEvaluator"
 default = "peagen.plugins.evaluator_pools.default:DefaultEvaluatorPool"
 
 [tool.setuptools.package-data]
-"peagen.schemas" = ["*.json", "extras/*.json"]
+"peagen.jsonschemas" = ["*.json", "extras/*.json"]
 "peagen" = [
   "tests/examples/locking_demo/*",
   "tests/examples/peagen_tomls/*",
